[Update] Added IBL + Fixed issues with memory
- Added some additional things to weakref class
- Uniform buffers are now created by the material class
- Added BufferPointer class (TODO: Should be changed to Buffer, and the other Buffer class to StorageBuffer)
- Added frustum culling (TODO: Maybe change the algorithm into a more efficient one)
- Added full pbr(all compute shaders)
- Fixed the texture cubemap mip map generation

+++ Frost/src/Frost/Platform/Vulkan/RayTracing/VulkanAccelerationStructure.cpp
@@ -198,7 +198,9 @@ namespace Frost
 
 			// Get the size result back
 			Vector<VkDeviceSize> compactSize(1);
-			vkGetQueryPoolResults(device, queryPool, 0, compactSize.size(), compactSize.size() * sizeof(VkDeviceSize), compactSize.data(), sizeof(VkDeviceSize), VK_QUERY_RESULT_WAIT_BIT);
+			vkGetQueryPoolResults(device, queryPool, 0,
+								  static_cast<uint32_t>(compactSize.size()), static_cast<uint32_t>(compactSize.size() * sizeof(VkDeviceSize)), compactSize.data(),
+								  sizeof(VkDeviceSize), VK_QUERY_RESULT_WAIT_BIT);
 
 
 			// Compacting
